Check for available hosts before creating a virtual machine
